Make it available as a component. Show it as the helper menus.
