+++ packages/tldraw/src/index.ts
@@ -78,6 +78,7 @@ export { useDefaultColorTheme } from './lib/shapes/shared/useDefaultColorTheme'
 export { useEditableText } from './lib/shapes/shared/useEditableText'
 export { TextShapeTool } from './lib/shapes/text/TextShapeTool'
 export { TextShapeUtil } from './lib/shapes/text/TextShapeUtil'
+export { Timer, type TimerProps } from './lib/shapes/timer/Timer'
 export { TimerShapeTool } from './lib/shapes/timer/TimerShapeTool'
 export { TimerShapeUtil } from './lib/shapes/timer/TimerShapeUtil'
 export { VideoShapeUtil } from './lib/shapes/video/VideoShapeUtil'
